LibPDF: Implement draw_gouraud_triangle()
This lets us paint shading types 4 and 5.

A type 4 or type 5 shading consists of a bunch of triangles
that usually share edges. For rendering these well, we need
a "watertight" triangle renderer, meaning that if two triangles
that share an edge are painted, the shared edge is ideally painted
exactly once.

This means we can't just use `Painter::fill_path()`: That will
give each triangle nice anti-aliased edges, but the blending
mode it uses will result in visible seams at the triangle edges.

Options for getting this right:

1. Disable anti-aliasing. This results in jaggy outer triangle edges.
   (But shadings are usually used with a clip mask and the outer
   edges usually aren't visible, so that's probably fine.)

2. Use coverage-to-alpha antialiasing, but draw the shading to
   a temporary, all-0 bitmap, and use a "plus" blending operator
   where all (alpha-premultiplied) channels are added to the
   destination color. For pixels that are on the shared edge,
   pixels are painted twice, but their coverage will add up to
   100%. Pixels on the outer edge will have nice coverage-as-alpha.
   Blending this temporary bitmap will produce nice anti-aliased
   outer edges and solid inner edges. It costs an additional buffer
   and blend. (...but shadings are usually used with a clip mask,
   and once we implement that, we'll use a temporary buffer and
   a blend for the clip mask already, and we could probably
   combine the clip and the shading buffer and blend.)

For now, this goes with option (1). (That's also what 3D APIs
usually do.)

In addition to drawing watertight triangles, we also need to
be able to associate a color with each corner, interpolate
that color in an arbitrary PDF color space (per spec
"Color Space: Special Considerations", only required for
the CIE-based color spaces and optional for the others,
but might as well get it right), and optionally apply
a PDF function per pixel on the interpolated value.

As for how to implement the fill, we have four options:

1. EdgeFlagPathRasterizer::fill_path() with a custom PaintStyle.
   The custom paint style computes barycentric coordinates per pixel
   and then interpolates colors based on that.
   + This approach could do the fancier antialias option (2) above
     with not too many changes
   + It handles arbitrary color spaces
   + It supports subpixels
   - It's likely fairly slow (it computes coverage data that
     we end up not using, etc)
   - We need to add a toggle to disable antialiasing
     (done in the previous commit)

2. LibGL. Gouraud triangles are a very "90s fixed function pipeline"
   3D API thing, so this seems fairly natural.
   + It'd be cool to use our OpenGL implementation!
   0 Supporting arbitrary functions would require converting the
     function to a 1D texture
   - It'd need shader support for arbitrary color space interpolation

3. Painter::draw_triangle() can draw triangles. We could teach it to
   associate a color with each corner.
   - Supporting arbitrary functions would be fairly ad-hoc
   - Doesn't do subpixels

4. A basic aliased triangle rasterizer with gouraud support doesn't
   need all that much code; we could have a custom rasterizer in
   Shading.cpp, based on the resources mentioned e.g. in
   https://github.com/aras-p/playdate-dither3d/blob/main/src/rasterizer_halfspace.h
   + Would be fun to write
   - Lacks synergy :P

This uses option 1 for now.