LibPDF: Use constant alpha for painting images
We're supposed to also use this for paths, texts, and shadings,
but we don't do that part yet. We're also still missing support
for most other things in the transparent imaging model (using
a transparency group as the graphic's state SMask, support for
transparency groups at all, blend modes, etc.) But this is a
small first step towards more support.