LibPDF: Add path clip in end_path_paint(), not in W / W* handler
W / W* just set a "add next path as clip" flag (which is *not*
stored in the graphics state), which is then used after a path
is painted.

Improves the appearance of the pattern in the upper right of
Tests/LibPDF/clip.pdf to almost match Acrobat Reader and PDFium.
(Firefox gets it mostly right; Preview.app is a bit confused.)