LibPDF: Use enclosing_int_rect() of clip path as bounding box
Previously, we used the rounded-to-int version.

When clipping the the clipping path's bounding box, neither seems
obviously better -- either you clip too little or too much.

But when we add support for clipping to full paths by rasterizing
the path to a temporary bitmap that's used as alpha channel, the
enclosing_int_rect() version will have the correct size for that
bitmap's dimensions, so switch to that.