Tests/LibGfx: Add missing EndOfPage segment to two jbig2 files
I had extracted white_47x23.jbig2 and black_47x23.jbig2 from a PDF
(and then tweaked them). PDFs store jbig2 data in the embedded
organization. In PDFs, the EndOfPage segment is stripped.

But in standalone files in the sequential organization, the EndOfPage
segment is required per spec (see 7.4.9 End of page segment syntax:
"Each page must have exactly one end of page segment associated with
it.").

I appended the bytes

    00 00 00 01 31 00 01 00 00 00 00

to each file. That's an EndOfPage segment with segment number 1,
associated with page 1. This should make the files valid.

(EndOfFile is only required for files in the random organization;
it's optional in the sequential organization from what I understand.)