LibGfx/JBIG2: Introduce a SymbolContexts struct
That's how all the arithmetic contexts for the generic region,
refinement region, and text region procedures are organized.
For the symbol dictionary procedure, this isn't necessary since
we don't need to pass the contexts in or share them with other
decoding procedures, but it makes things more consistent and
the code looks a bit nicer this way too.

Also add a missing comment about the "IAEX" name while here.

No behavior change.