LibGfx/JBIG2: Read and log segment retention flags
I think the way retention flags are supposed to work is that they
tell you how long segments have to be kept in memory. I imagine
that a symbol dictionary segment will have its flag set until the
last text segment referring to it, which will then set its flag
to false.

For now, we do nothing with these, but if files in the wild do
behave this way, we could add at least a check that segments
only go from retained to non-retained once and then stay
non-retained.