LibGUI: Don't render trailing whitespace when cursor is at last column
Previously it was pretty annoying to type on programs like TextEditor
because after every space the trailing whitespace visualizer kicked in
and a red square showed up right where the cursor was.

Now it's only rendered when the cursor is not at the end of the line,
which makes typing feel better!