Utilities+LibGfx: Add jbig2-from-json
jbig2-from-json can create JBIG2 files from json descriptions.
The intent is to use it to create test files.

jbig2-from-json only allows producing valid JBIG2 files.
While producing invalid files might make sense for a test file creator,
a fuzzer is better for testing invalid files and the risk of
accidentally creating an invalid file is high otherwise.  This makes it
possible to not put a ton of validity checks in the writer -- just
creating the file in memory and checking if our strict loader can read
it is sufficient.

JBIG2Writer grows a function for creating a JBIG2 files from a
detailed description of the output. For now, it can only create
PageInformation and EndOfPage segments. Once it learns more
segment types, the existing JBIG2Writer::encode() function can
be implemented on top of this new API.

jbig2-from-json then transforms JSON to this internal structure
and calls the new API.