LibGfx/JBIG2+Tests: Add another spec comment
Adds a spec comment for the case from #26132.

Also add a test for the situation from that PR: The very first (and in
the case of a one-stripe image, only) stripe is allowed to be one pixel
higher than maximum_stripe_size: For e.g. a one-stripe 400-pixel high
image, the first and only EndOfStripe y_coordinate will be 399, and due
to spec wording maximum_stripe_size can be 399 as well.