Reapply "LibPDF: Add basic tiled, coloured pattern rendering"
This reverts commit 6032c06f6b293bcb46650989d678751af47ee977.

This relands the original commit with fixups to make `PatternColorSpace`
more robust to unexpected inputs.

While working on #26292, I noticed that while fills/strokes can be
Gfx::PaintStyles in LibPDF, currently they never are, which confused me.

It turns out that a few years ago, tiled patterns were implemented in
PR #22197, but were partially reverted due to crashes in #22364, leaving
around a partial implementation.

It appears the author was going to look into the crashes, but never got
around to it. So, this patch resolves all the crashes on `0000.zip`,
mainly by checking types and not assuming dictionary keys exist.

Summary of `./Meta/test_pdf.py ~/Downloads/0000`:

```
Stacks:
0 crashes (0.0%)
0 distinct crash stacks

7 failed to open (0.7%)
    /home/macdue/Downloads/0000/0000346.pdf
    /home/macdue/Downloads/0000/0000202.pdf
    /home/macdue/Downloads/0000/0000399.pdf
    /home/macdue/Downloads/0000/0000421.pdf
    /home/macdue/Downloads/0000/0000480.pdf
    /home/macdue/Downloads/0000/0000920.pdf
    /home/macdue/Downloads/0000/0000819.pdf

3 files with password (0.3%)

909 files without issues (90.9%)
```

Co-authored-by: MacDue <macdue@dueutil.tech>