Tests/LibGfx: Tweak huffman text region refinement custom table test
The custom tables in this test are meant to be the default tables with
some rows swapped, but I had a typo in the tables that are meant to
be B.1 with rows swapped: B.1 has a lowest value of 0, not 1.

No real behavior change (the test checked what it was supposed to
check with the "wrong" value too), just pedantry.