Kernel/DeviceTree: Add a new interrupt translation API
This new API can be used by devicetree drivers to get the (translated)
interrupt numbers for the interrupts of a device node.

Currently, the translation of interrupt specifiers to interrupt numbers
is implemented inline in each driver. This is not really scalable, since
the interrupt specifier format depends on the interrupt controller the
interrupt is connected to, so each driver would need to handle all
interrupt specifier formats.