Tests/LibGfx: Let jbig2/json/compile.sh process files in parallel
This uses `xargs -P` as a fairly simple parallelism mechanism.  It
relies on `echo` being atomic in practice (for small enough outputs).

(Alternatively, each child could write to a dedicated log file in
some temp dir, and we could cat all the log files at the end. For
now, this seems like it's good enough.)

Takes the time to run compile.sh from 1.65s to 0.23s on my system.