LibGfx/ICC: Add a built-in LAB mft2 profile
This is even weirder than the XYZ mft2 profile added in an
earlier commit in this PR:

mft2 uses a "legacy 16-bit PCSLAB encoding" instead of the "normal"
16-bit ICC PCSLAB encoding, and this being an identity mapping leaks
this into the outputs. We again linearly map this internal 16-bit
encoding to 8 bits.

This legacy encoding maps 0 to 0.0 and FF00 to 1.0 for L
(and FFFF to a bit more than 1.0), and 0 to -128.0 and FF00 to
127.0 for a* and b* (and FFFF to a bit more than 127.0).

This means this produces different values than our built-in LAB mft1
profile:

    % echo 'pcslab(100, 127, -128)' | \
          icc -n LAB_mft2 --stdin-u8-from-pcs | \
          icc -n LAB --stdin-u8-to-pcs
    pcslab(99.60784, 126, -128)

Staying in LAB for both steps is exact, of course:

    % echo 'pcslab(100, 127, -128)' | \
          icc -n LAB --stdin-u8-from-pcs | \
          icc -n LAB --stdin-u8-to-pcs
    pcslab(100, 127, -128)

Staying in LAB_mft2 for both steps is also exact, up to u8
resolution (100 and 127 don't exactly map to an u8 value with
our "16-bit legacy PCSLAB encoding mapped to u8" encoding):

    % echo 'pcslab(100, 127, -128)' | \
          icc -n LAB_mft2 --stdin-u8-from-pcs | \
          icc -n LAB_mft2 --stdin-u8-to-pcs
    pcslab(99.99693, 126.99219, -128)

This is just because the same PCSLAB value is encoded slightly
differently in the mft1 and mft2 LAB encodings:

    % echo 'pcslab(100, 127, -128)' | \
          icc -n LAB --stdin-u8-from-pcs
    255, 255, 0

    % echo 'pcslab(100, 127, -128)' | \
          icc -n LAB_mft2 --stdin-u8-from-pcs
    254, 254, 0

(There's no XYZ mft1 encoding, so at least we don't have mft1 and mft2
profiles that have different output.)

Use this profile to add a roundtrip test for Lut16TagData when PCS
is PCSLAB.