LibGfx/JBIG2Loader+jbig2-from-json: Make JBIG2Loader strictness settable
Clients of JBIG2Loader can now pick how strictly invalid images will
be rejected.

By default, it's set to permissive, which allows spec violations that
happen in practice and are easy to work around. At the moment, this
is just for images in the Power JBIG2 suite, but this will be used to
relax some other checks soon.

jbig2-from-json now requests strict spec-compliant checking. This for
now has the effect that manually adding a comment advertising a file
as being a Power JBIG2 file will no longer allow a few quirks when
using jbig2-from-json.