Tests/LibGfx: Add huffman-coded version of symbolrefine-textrefine.jbig2
This test is identical to bitmap-symbol-symbolrefine-textrefine.json,
except with

    "uses_huffman_encoding": true,

in the "flags" section of seach segment.

While the arithmetically-coded version of this file decodes fine
everywhere, the huffman-coded version decodes only correctly in
Acrobat Reader, and almost correctly in Preview.app. PDFium,
pdf.js, and mutool can't currently decode it.